Validate ImageDeletedEventFunction

+++ Messaging/Features/ImageDeletedEventFunction.cs
@@ -1,9 +1,11 @@
 using Azure.Storage.Blobs;
 using Couple.Messaging.Data;
 using Couple.Messaging.Model;
+using FluentValidation;
 using Microsoft.Azure.WebJobs;
 using Microsoft.Azure.WebJobs.Extensions.EventGrid;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Logging;
 
 namespace Couple.Messaging.Features;
 
@@ -17,8 +19,18 @@ public class ImageDeletedEventFunction
     }
 
     [FunctionName("ImageDeletedEventFunction")]
-    public async Task Run([EventGridTrigger] Event @event)
+    public async Task Run([EventGridTrigger] Event @event, ILogger log)
     {
+        var validationResult = new Validator().Validate(@event);
+        if (!validationResult.IsValid)
+        {
+            var errorMessage = validationResult.Errors
+                .Select(error => error.ToString())
+                .ToString();
+            log.LogWarning("{ErrorMessage}", errorMessage);
+            return;
+        }
+
         var connectionString = Environment.GetEnvironmentVariable("ImagesConnectionString");
         var client = new BlobClient(connectionString, "images", @event.Subject);
         await client.DeleteIfExistsAsync();
@@ -31,6 +43,14 @@ public class ImageDeletedEventFunction
             _context.Images
                 .Remove(toDelete);
             await _context.SaveChangesAsync();
+        }
+    }
+
+    public class Validator : AbstractValidator<Event>
+    {
+        public Validator()
+        {
+            RuleFor(dto => dto.Subject).NotEmpty();
         }
     }
 }
